+++ src/components/ModalLogout/ModalLogout.js
@@ -2,14 +2,16 @@
 /* eslint-disable react/button-has-type */
 /* eslint-disable no-unused-vars */
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { connect } from 'react-redux';
 import styleModalLogout from './ModalLogout.module.css';
 import ModalBackdrop from '../ModalBackdrop/ModalBackdrop';
+import { ModalLogoutClosed } from '../../redux/global/globalActions';
+import { logout } from '../../redux/auth/authOperation';
 
-const ModalLogout = ({ onClose }) => {
+const ModalLogout = ({ onOpenModalLogout, onLogout }) => {
   return (
     <>
-      <ModalBackdrop onClose={onClose}>
+      <ModalBackdrop onClose={onOpenModalLogout}>
         <div className={styleModalLogout.modalSize}>
           <h2 className={styleModalLogout.modalTitleText}>
             Ви дійсно бажаєте вийти?
@@ -17,11 +19,17 @@ const ModalLogout = ({ onClose }) => {
 
           <div className={styleModalLogout.prizeContainer} />
           <div className={styleModalLogout.buttonContainer}>
-            <button className={styleModalLogout.point_amount}>
-              <p className={styleModalLogout.point_amount_p}>Так</p>
+            <button
+              className={styleModalLogout.point_amount}
+              onClick={onLogout}
+            >
+              Так
             </button>
-            <button onClick={onClose} className={styleModalLogout.point_amount}>
-              <p className={styleModalLogout.point_amount_p}>Ні</p>
+            <button
+              onClick={onOpenModalLogout}
+              className={styleModalLogout.point_amount}
+            >
+              Ні
             </button>
           </div>
         </div>
@@ -30,4 +38,9 @@ const ModalLogout = ({ onClose }) => {
   );
 };
 
-export default ModalLogout;
+const mapDispatchToProps = dispatch => ({
+  onOpenModalLogout: () => dispatch(ModalLogoutClosed()),
+  onLogout: () => dispatch(logout()),
+});
+
+export default connect(null, mapDispatchToProps)(ModalLogout);
